Updated keshlib, reworked the demo view controller to adopt the new lib functions, iOS9 and XCode 7 adjustments (UIAlertView gives a warning now in iOS9, but to keep compatibility with iOS6 we don't use the new UIAlertController class)

+++ KSKeshLibraryExt.framework/Versions/A/Headers/KSLimitUsage.h
@@ -27,7 +27,7 @@ typedef enum LimitType {
 @property(copy,readonly) KSAmount *maxAmount;
 
 - (id)initWithLimitType:(NSString*)limitType currentAmount:(KSAmount *)currentAmount maxAmount:(KSAmount *)maxAmount;
-- (float)percentage;
+- (double)percentage;
 - (NSString *)actualString;
 - (NSString *)maximumString;
 - (NSString*)remainingString;
